pyproject.toml
[tool.poetry]
name = "ariadne-utils"
version = "0.1.0"
description = ""
authors = ["Marc Ford <mrfxyz567@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/mfdeux/ariadne-utils"
packages = [{ include = "ariadne_utils" }]
include = ["data/*"]

[tool.poetry.dependencies]
python = "^3.8"
ariadne = "^0.12.0"
click = "^7.1.2"
python-dateutil = "^2.8.1"
validator-collection = "^1.4.2"
pytimeparse = "^1.1.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
isort = "^5.3.0"

[tool.poetry.scripts]
ariadne-utils = "ariadne_utils.cli:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target_version = ['py38']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''

[tool.isort]
line_length = 88
indent = '    '
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
